Migrate openai_usages
